remove inefficient .find for every grid day
