+++ components/ActivityGrid.tsx
@@ -1,13 +1,9 @@
-import { ReactNode, useEffect, useState } from "react";
-import { addDays, format, subDays } from "date-fns";
+import { ReactNode } from "react";
+import { format } from "date-fns";
 
 export interface ActivityDay {
-    date: string,
-    count: number,
-}
-
-interface GridDay {
     date: Date,
+    count: number,
     day: number,
     week: number,
 }
@@ -23,35 +19,10 @@ const GridLabel = ({ row, col, children }: { row: number, col: number, children:
     ><span>{children}</span></div>
 )
 
-export default function ActivityGrid({ data, label, color, endDate = new Date() }: { data: ActivityDay[], label?: string, color?: string, endDate?: Date }) {
-    const [gridDays, setGridDays] = useState<GridDay[]>([]);
+export default function ActivityGrid({ data, label, color }: { data: ActivityDay[], label?: string, color?: string }) {
     const numCols = 53;
 
-    useEffect(() => {
-        const today = endDate;
-        const todayDayOfWeek = today.getDay();
-        const todayFirstDayOfWeek = subDays(today, todayDayOfWeek);
-        const firstDayOnGraph = subDays(todayFirstDayOfWeek, (numCols - 1) * 7);
-
-        let newGridDays: GridDay[] = [];
-        let currDay = firstDayOnGraph;
-        let week = 0;
-
-        while (currDay <= today) {
-            newGridDays.push({
-                date: currDay,
-                day: currDay.getDay(),
-                week: week,
-            });
-
-            if (currDay.getDay() === 6) week++;
-            currDay = addDays(currDay, 1);
-        }
-
-        setGridDays(newGridDays);
-    }, [data])
-
-    const monthChangeDays = gridDays.filter((d, i, a) => (
+    const monthChangeDays = data.filter((d, i, a) => (
         i === 0 || a[i - 1].date.getMonth() !== d.date.getMonth()
     ));
 
@@ -76,24 +47,20 @@ export default function ActivityGrid({ data, label, color, endDate = new Date() 
             <GridLabel row={3} col={1}>Mon</GridLabel>
             <GridLabel row={5} col={1}>Wed</GridLabel>
             <GridLabel row={7} col={1}>Fri</GridLabel>
-            {gridDays.map(d => {
-                const thisDataPoint = data.find(x => x.date === format(d.date, "yyyy-MM-dd"));
-
-                return (
-                    <div
-                        style={{
-                            backgroundColor: thisDataPoint ? (color || "#0026ff") : "#000",
-                            opacity: thisDataPoint ? thisDataPoint.count / maxCount : 0.05,
-                            width: 14,
-                            height: 14,
-                            gridRow: d.day + 2,
-                            gridColumn: d.week + 2,
-                            borderRadius: 3,
-                        }}
-                        key={format(d.date, "yyyy-MM-dd")}
-                    />
-                );
-            })}
+            {data.map(d => (
+                <div
+                    style={{
+                        backgroundColor: d.count > 0 ? (color || "#0026ff") : "#000",
+                        opacity: d.count > 0 ? d.count / maxCount : 0.05,
+                        width: 14,
+                        height: 14,
+                        gridRow: d.day + 2,
+                        gridColumn: d.week + 2,
+                        borderRadius: 3,
+                    }}
+                    key={format(d.date, "yyyy-MM-dd")}
+                />
+            ))}
         </div>
     );
 }
